Clear up 1 more clang++ warning

+++ boost/multiprecision/cpp_int/import_export.hpp
@@ -228,7 +228,7 @@ OutputIterator export_bits(
       ++chunks;
 
          std::ptrdiff_t bit_location = msv_first ? static_cast<std::ptrdiff_t>(bitcount - chunk_size) : 0;
-   const std::ptrdiff_t bit_step     = msv_first ? -static_cast<int>(chunk_size) : chunk_size;
+   const std::ptrdiff_t bit_step     = msv_first ? static_cast<std::ptrdiff_t>(-static_cast<std::ptrdiff_t>(chunk_size)) : static_cast<std::ptrdiff_t>(chunk_size);
    while (bit_location % bit_step)
       ++bit_location;
